do not set publish state to default on an image refresh metadata LensInfo for exif.php

+++ zp-core/admin-functions.php
@@ -2181,7 +2181,7 @@ function processAlbumEdit($index, $album, &$redirectto) {
 	$album->setDateTime(sanitize($_POST[$prefix."albumdate"]));
 	$album->setLocation(process_language_string_save($prefix.'albumlocation', 3));
 	if (isset($_POST[$prefix.'thumb'])) $album->setAlbumThumb(sanitize($_POST[$prefix.'thumb']));
-	$album->setShow(isset($_POST[$prefix.'Published']));
+	$album->setShow((int) isset($_POST[$prefix.'Published']));
 	$album->setCommentsAllowed(isset($_POST[$prefix.'allowcomments']));
 	$sorttype = strtolower(sanitize($_POST[$prefix.'sortby'], 3));
 	if ($sorttype == 'custom') {
